Move clear scenery tool functions from TopToolbar.cpp into ClearScenery
